+++ splunkbase/releasenotes.md
@@ -1,5 +1,11 @@
 # Cisco Intersight Add-on for Splunk Release Notes
 
+## 1.3.1 - 3 August 2022
+
+- Found a serious bug in checkpointing that might cause Audit Logs and Alerts to never be retrieved for an input
+- Additional pruning for cisco:intersight:computePhysicalSummaries and a few other items
+- Added cisco:intersight:equipmentChassis to Compute inventory
+
 ## 1.3.0 - 28 July 2022
 
 >New Inventory types are not automatically added to existing configurations.  Please review your input configurations after upgrading to enable the new options.